feat: Add negated matchers for include and be_within
The code changes include defining negated matchers for the `include` and `be_within` matchers in the `arrays_spec.rb` and `be_within_spec.rb` files respectively. These negated matchers allow for checking if an element is not included in an array and if a number is not within a certain range. This addition enhances the flexibility and expressiveness of the test assertions.